fixed package publishing for private packages
