+++ packages/beachball/src/PackageInfo.ts
@@ -1,0 +1,10 @@
+export interface PackageInfo {
+  name: string;
+  packageJsonPath: string;
+  version: string;
+  dependencies?: { [dep: string]: string };
+  devDependencies?: { [dep: string]: string };
+  disallowedChangeTypes: string[];
+  defaultNpmTag: string;
+  private: boolean;
+}
