+++ packages/beachball/src/fixtures/package.ts
@@ -1,17 +1,17 @@
 import fs from 'fs';
 import path from 'path';
 import * as tmp from 'tmp';
-import { PackageInfo } from '../monorepo';
+import { PackageInfo } from '../PackageInfo';
 
 export const testTag = 'testbeachballtag';
 
 const testPackage = {
-  name: "testbeachballpackage",
-  version: "0.6.0"
+  name: 'testbeachballpackage',
+  version: '0.6.0',
 };
 
 // Create a test package.json in a temporary location for use in tests.
-var tmpPackageFile = path.join(tmp.dirSync().name, 'package.json')
+var tmpPackageFile = path.join(tmp.dirSync().name, 'package.json');
 
 var testPackageJson = JSON.stringify(testPackage);
 fs.writeFileSync(tmpPackageFile, testPackageJson, 'utf8');
@@ -22,5 +22,5 @@ export const testPackageInfo: PackageInfo = {
   defaultNpmTag: 'latest',
   version: testPackage.version,
   disallowedChangeTypes: [],
-  private: false
+  private: false,
 };
